Initial pass at refactoring stats_table.py to use query builders
