+++ posthog/hogql_queries/web_analytics/query_builders/__init__.py
@@ -1,0 +1,19 @@
+from posthog.hogql_queries.web_analytics.query_builders.base import BaseStatsTableQueryBuilder
+from posthog.hogql_queries.web_analytics.query_builders.base_bounce_query_builder import BaseBounceQueryBuilder
+from posthog.hogql_queries.web_analytics.query_builders.frustration_metrics_query_builder import (
+    FrustrationMetricsQueryBuilder,
+)
+from posthog.hogql_queries.web_analytics.query_builders.main_query_builder import MainQueryBuilder
+from posthog.hogql_queries.web_analytics.query_builders.path_bounce_query_builder import PathBounceQueryBuilder
+from posthog.hogql_queries.web_analytics.query_builders.path_scroll_bounce_query_builder import (
+    PathScrollBounceQueryBuilder,
+)
+
+__all__ = [
+    "BaseStatsTableQueryBuilder",
+    "BaseBounceQueryBuilder",
+    "MainQueryBuilder",
+    "PathBounceQueryBuilder",
+    "PathScrollBounceQueryBuilder",
+    "FrustrationMetricsQueryBuilder",
+]
